refactor(device): migrate DeviceValidator to DeviceGuard pattern
Migrate DeviceValidator to DeviceGuard applying the guard clause pattern
to reduce ambiguity.
